Make t1b1 generic over balanced and unbalanced ternary
This commit also removes certain integration tests that check for conversion between ternary buffers of one encoding containing Utrits to buffers of another encoding containing Btrits. The changes to RawEncoding, RawEncodingBuf, Trits, and TritsBuf prevent this kind of conversion. This is considered reasonable because a conversion between encodings, say t1b1 to t3b1, should be orthogonal to an integer representation